requirements.txt
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
idna==2.8
pycparser==2.19
python-mpd2==1.0.0
requests==2.21.0
spotipy==2.4.4
urllib3==1.24.2
